requirements.txt
selenium==4.26.1
pillow==10.4.0
python-dotenv==1.0.0
requests==2.31.0
selenium==4.26.1
swarms==6.9.6
swarm_models==0.3.1
moviepy==2.1.2
# Add other dependencies your application needs 